Add files via upload
A 500,000 data set of previous transactions which represent orders across different regions was been provided for the 8 teams of Professionals in the 2021 utiva data hackathon. I collaborated with my team to team to derive analytics and intelligence using excellent visualization tools to represent our insights and made a presentation to the leadership team where my team finally emerged the winner of the 11th December 2021 Utiva Data School Hackathon.